fixup! Don't exclude quite so much in a no-sock build
exclude even less and move the socket dependency into bio_wait() in bio_lib.c
