+++ include/openssl/bio.h
@@ -662,9 +662,9 @@ int BIO_dgram_sctp_msg_waiting(BIO *b);
 int BIO_sock_should_retry(int i);
 int BIO_sock_non_fatal_error(int error);
 int BIO_socket_wait(int fd, int for_read, time_t max_time);
-int BIO_wait(BIO *bio, time_t max_time);
-int BIO_connect_retry(BIO *bio, int timeout);
 # endif
+int BIO_wait(BIO *bio, time_t max_time, unsigned long milliseconds);
+int BIO_connect_retry(BIO *bio, int timeout);
 
 int BIO_fd_should_retry(int i);
 int BIO_fd_non_fatal_error(int error);
